fixes for make and Abstract driven by refs errors on 21.sasl test case

+++ sasl/reduce.c
@@ -61,7 +61,7 @@ static int tag_reductions[TagCount];
 static void indent(FILE *where, long int n)
 {
   static char b[] = "                ";
-  static int max;
+  static long int max;
 
   max = strlen(b);
